requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.5
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.0
psycopg2==2.9.1
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.11
ruamel.yaml.clib==0.2.6
simplejson==3.17.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
